refactor(bot): rename CogBase to BaseCog
