+++ fazcord/bot/cog/info_cog.py
@@ -1,4 +1,4 @@
-from fazcord.bot.cog._cog_base import CogBase
+from fazcord.bot.cog._base_cog import CogBase
 
 
 class InfoCog(CogBase): ...
