+++ fazcord/bot/cog/wynn_analyze_cog.py
@@ -1,4 +1,4 @@
-from fazcord.bot.cog._cog_base import CogBase
+from fazcord.bot.cog._base_cog import CogBase
 
 
 class WynnAnalyzeCog(CogBase): ...
